% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageROIs.R
\name{averageROIs}
\alias{averageROIs}
\title{Average ROIs}
\usage{
averageROIs(
  data_dt,
  sampleCol_v = "Sample_ID",
  prefix_v = "",
  split_v = "ROI",
  splitLabs_v = c("Patient", "Sample"),
  by_v = "Patient",
  metaCols_v = NULL,
  meta_dt = NULL,
  mergeCol_v = NULL
)
}
\arguments{
\item{data_dt}{table of multiplex results. Can be original counts (StudyCounts) or densities calculated from those}

\item{sampleCol_v}{name of column that contains the sample IDs to use for aggregation}

\item{prefix_v}{character vector that is prefix of values in sampleCol_v. Default is blank (no prefix). Whatever
is provided will be erased from the sample id.}

\item{split_v}{character vector of what to split sampleCol_v on in order to separate out patient from samples}

\item{splitLabs_v}{character vector containing names of what to call results of sampleCol_v splitting}

\item{by_v}{what to aggregate by. Default is Patient, but could be treatment or something if included in the name}

\item{metaCols_v}{columns to exclude from aggregation}
}
\value{
data.table with same columns. Rows are now summarized by by_v.
}
\description{
Calculate average values for ROIs originating from same sample (i.e. patient+timepoint)
}
