% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDensity.R
\name{calcDensity}
\alias{calcDensity}
\title{Calculate Density}
\usage{
calcDensity(
  data_dt,
  meta_dt = NULL,
  areaCol_v = "Area",
  idCol_v = "Sample_ID",
  metaCols_v = NULL
)
}
\arguments{
\item{data_dt}{data output by gating. File form is usually "StudyCounts"}

\item{meta_dt}{metadata file containing the area column}

\item{areaCol_v}{column to use as area for density calculations. Default is 'Area'}

\item{idCol_v}{column in both data_dt and meta_dt that identifies samples}

\item{metaCols_v}{columns to ignore during density calculation (optional)}
}
\value{
data.table that is same as input except each cell is a density instead of a count. Also 'Area' column removed
}
\description{
Convert mIHC cell counts into cell densities. Current format is for the area column to be recorded
in the metadata file. Sometimes it's in the data file. This should be able to handle either.
}
