% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionalAverageROIs.R
\name{functionalAverageROIs}
\alias{functionalAverageROIs}
\title{Average Functional ROIs}
\usage{
functionalAverageROIs(
  functional_lsdt,
  id_v,
  col_v = "class",
  meta_dt = NULL,
  metaCol_v = NULL,
  sampleCol_v = "Sample_ID"
)
}
\arguments{
\item{functional_lsdt}{list of functional tables (either counts or density), one for each patient+timepoint+ROI}

\item{id_v}{sample ID that will be averaged over.}

\item{col_v}{column name in each functional data.table that is the identifier (i.e. not a functional value)}
}
\value{
data.table containing the averaged results of id_v
}
\description{
Calculate average functional values for ROIs originating from same sample (i.e. patient+timepoint)
}
