% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mIHC-package.R
\docType{package}
\name{mIHC-package}
\alias{mIHC}
\alias{mIHC-package}
\title{mIHC: Multiplex IHC Functions}
\description{
Collection of plotting and data wrangling functions for use with mIHC reports.
}
\author{
\strong{Maintainer}: Wes Horton \email{hortowe@ohsu.edu}

}
\keyword{internal}
