\name{phy}
\alias{phy}
\docType{data}
\title{
An example microbiome and meta data
}
\description{
This is an example data in the phyloseq format.
}
\usage{data("phy")}
\format{
phyloseq
}
\references{
McMurdie and Holmes (2013). phyloseq: An R package for reproducible interactive analysis and graphics of
microbiome census data. PLoS ONE. 8(4):e61217

}
\author{
Hyunwook Koh
}
\examples{
require(phyloseq)

data(phy)
phy

}
