% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCSVOrExcel.R
\name{readCSVOrExcel}
\alias{readCSVOrExcel}
\title{Read csv of excel}
\usage{
readCSVOrExcel(file_v)
}
\arguments{
\item{file_v}{path to file to be read in}
}
\value{
data.table
}
\description{
Read in csv file or first worksheet of an excel workbook
}
